# Allocate the initial generator buffer on the stack