--- a/ext/json/ext/parser/parser.c
+++ b/ext/json/ext/parser/parser.c
@@ -2159,7 +2159,7 @@ static VALUE cJSON_parser_s_allocate(VALUE klass)
 {
     JSON_Parser *json;
     VALUE obj = TypedData_Make_Struct(klass, JSON_Parser, &JSON_Parser_type, json);
-    fbuffer_init(&json->fbuffer, 0);
+    fbuffer_stack_init(&json->fbuffer, 0, NULL, 0);
     return obj;
 }