@casperisfine 2024-10-29
Interestingly, this change is severely degrading the performance on larger benchmarks:

Before:

```
== Encoding twitter.json (466906 bytes)
ruby 3.3.4 (2024-07-09 revision be1089c8ec) +YJIT [arm64-darwin23]
Warming up --------------------------------------
                json   203.000 i/100ms
                  oj   225.000 i/100ms
Calculating -------------------------------------
                json      2.105k (± 1.7%) i/s  (475.03 μs/i) -     10.556k in   5.015961s
                  oj      2.253k (± 5.5%) i/s  (443.90 μs/i) -     11.250k in   5.015686s

Comparison:
                json:     2105.1 i/s
                  oj:     2252.7 i/s - same-ish: difference falls within error


== Encoding citm_catalog.json (500298 bytes)
ruby 3.3.4 (2024-07-09 revision be1089c8ec) +YJIT [arm64-darwin23]
Warming up --------------------------------------
                json   133.000 i/100ms
                  oj   122.000 i/100ms
Calculating -------------------------------------
                json      1.349k (± 1.9%) i/s  (741.16 μs/i) -      6.783k in   5.029134s
                  oj      1.277k (± 2.0%) i/s  (783.34 μs/i) -      6.466k in   5.067016s

Comparison:
                json:     1349.2 i/s
                  oj:     1276.6 i/s - 1.06x  slower
```

After:

```
== Encoding twitter.json (466906 bytes)
ruby 3.3.4 (2024-07-09 revision be1089c8ec) +YJIT [arm64-darwin23]
Warming up --------------------------------------
                json   160.000 i/100ms
                  oj   225.000 i/100ms
Calculating -------------------------------------
                json      1.577k (± 2.0%) i/s  (634.20 μs/i) -      8.000k in   5.075561s
                  oj      2.264k (± 2.3%) i/s  (441.79 μs/i) -     11.475k in   5.072205s

Comparison:
                json:     1576.8 i/s
                  oj:     2263.5 i/s - 1.44x  faster


== Encoding citm_catalog.json (500298 bytes)
ruby 3.3.4 (2024-07-09 revision be1089c8ec) +YJIT [arm64-darwin23]
Warming up --------------------------------------
                json   101.000 i/100ms
                  oj   123.000 i/100ms
Calculating -------------------------------------
                json      1.033k (± 2.6%) i/s  (968.06 μs/i) -      5.252k in   5.087617s
                  oj      1.257k (± 2.2%) i/s  (795.54 μs/i) -      6.396k in   5.090830s

Comparison:
                json:     1033.0 i/s
                  oj:     1257.0 i/s - 1.22x  faster
```

Need to figure out why before merging this.