@byroot 2024-10-29
Before:

<img width="1170" alt="Capture d’écran 2024-10-29 à 11 38 34" src="https://github.com/user-attachments/assets/9320edab-0c6f-469d-9a03-dde189fb1a95">

After:

<img width="1175" alt="Capture d’écran 2024-10-29 à 11 38 21" src="https://github.com/user-attachments/assets/4aff584e-a695-4ce1-83bd-a2999f15388d">

It's not super clear, but it looks like the callee of `generate_json_string` are no longer inlined?