@byroot 2024-10-30
I'm already seeing  that exact performance on `twitter.json` and `citm_catalog.json`, so I think we didn't actually degrade performance on larger benchmarks.